+++ src/modules/bank.c
@@ -164,8 +164,8 @@ static int bank_blit4(lua_State *L)
     LUAX_SIGNATURE_END
     Bank_Class_t *instance = (Bank_Class_t *)lua_touserdata(L, 1);
     int cell_id = lua_tointeger(L, 2);
-    double x = (double)lua_tonumber(L, 3);
-    double y = (double)lua_tonumber(L, 4);
+    float x = (float)lua_tonumber(L, 3);
+    float y = (float)lua_tonumber(L, 4);
 #ifdef __DEBUG_API_CALLS__
     Log_write(LOG_LEVELS_DEBUG, "Bank.blit() -> %d, %.f, %.f", cell_id, x, y);
 #endif
@@ -190,9 +190,9 @@ static int bank_blit5(lua_State *L)
     LUAX_SIGNATURE_END
     Bank_Class_t *instance = (Bank_Class_t *)lua_touserdata(L, 1);
     int cell_id = lua_tointeger(L, 2);
-    double x = (double)lua_tonumber(L, 3);
-    double y = (double)lua_tonumber(L, 4);
-    double rotation = (double)lua_tonumber(L, 5);
+    float x = (float)lua_tonumber(L, 3);
+    float y = (float)lua_tonumber(L, 4);
+    float rotation = (float)lua_tonumber(L, 5);
 #ifdef __DEBUG_API_CALLS__
     Log_write(LOG_LEVELS_DEBUG, "Bank.blit() -> %d, %.f, %.f, %.f", cell_id, x, y, rotation);
 #endif
@@ -218,10 +218,10 @@ static int bank_blit6(lua_State *L)
     LUAX_SIGNATURE_END
     Bank_Class_t *instance = (Bank_Class_t *)lua_touserdata(L, 1);
     int cell_id = lua_tointeger(L, 2);
-    double x = (double)lua_tonumber(L, 3);
-    double y = (double)lua_tonumber(L, 4);
-    double scale_x = (double)lua_tonumber(L, 5);
-    double scale_y = (double)lua_tonumber(L, 6);
+    float x = (float)lua_tonumber(L, 3);
+    float y = (float)lua_tonumber(L, 4);
+    float scale_x = (float)lua_tonumber(L, 5);
+    float scale_y = (float)lua_tonumber(L, 6);
 #ifdef __DEBUG_API_CALLS__
     Log_write(LOG_LEVELS_DEBUG, "Bank.blit() -> %d, %.f, %.f, %.f, %.f", cell_id, x, y, scale_x, scale_y);
 #endif
@@ -248,11 +248,11 @@ static int bank_blit7(lua_State *L)
     LUAX_SIGNATURE_END
     Bank_Class_t *instance = (Bank_Class_t *)lua_touserdata(L, 1);
     int cell_id = lua_tointeger(L, 2);
-    double x = (double)lua_tonumber(L, 3);
-    double y = (double)lua_tonumber(L, 4);
-    double scale_x = (double)lua_tonumber(L, 5);
-    double scale_y = (double)lua_tonumber(L, 6);
-    double rotation = (double)lua_tonumber(L, 7);
+    float x = (float)lua_tonumber(L, 3);
+    float y = (float)lua_tonumber(L, 4);
+    float scale_x = (float)lua_tonumber(L, 5);
+    float scale_y = (float)lua_tonumber(L, 6);
+    float rotation = (float)lua_tonumber(L, 7);
 #ifdef __DEBUG_API_CALLS__
     Log_write(LOG_LEVELS_DEBUG, "Bank.blit() -> %d, %.f, %.f, %.f, %.f, %.f", cell_id, x, y, scale_x, scale_y, rotation);
 #endif
@@ -281,13 +281,13 @@ static int bank_blit9(lua_State *L)
     LUAX_SIGNATURE_END
     Bank_Class_t *instance = (Bank_Class_t *)lua_touserdata(L, 1);
     int cell_id = lua_tointeger(L, 2);
-    double x = (double)lua_tonumber(L, 3);
-    double y = (double)lua_tonumber(L, 4);
-    double scale_x = (double)lua_tonumber(L, 5);
-    double scale_y = (double)lua_tonumber(L, 6);
-    double rotation = (double)lua_tonumber(L, 7);
-    double anchor_x = (double)lua_tonumber(L, 8);
-    double anchor_y = (double)lua_tonumber(L, 9);
+    float x = (float)lua_tonumber(L, 3);
+    float y = (float)lua_tonumber(L, 4);
+    float scale_x = (float)lua_tonumber(L, 5);
+    float scale_y = (float)lua_tonumber(L, 6);
+    float rotation = (float)lua_tonumber(L, 7);
+    float anchor_x = (float)lua_tonumber(L, 8);
+    float anchor_y = (float)lua_tonumber(L, 9);
 #ifdef __DEBUG_API_CALLS__
     Log_write(LOG_LEVELS_DEBUG, "Bank.blit() -> %d, %.f, %.f, %.f, %.f, %.f, %.f, %.f", cell_id, x, y, scale_x, scale_y, rotation, anchor_x, anchor_y);
 #endif
